Added a new CI pipeline creating a github release for all new tags on the master branch.
Derived from the `master` pipeline it replaces the unfiltered `src`
commit stream with a tag-filtered stream, and the `dist` task with
`release`.

The new task assembles a release from the distribution tarballs, using
the raw executables as assets, not the archives, and uploads the
result to github. It further bundles the assembled release into a
single tarball and uploads that to s3, under a name which will not
conflict with regular builds.

[trello#TirKTrmf]